fix pep8 format

+++ solvers/spams.py
@@ -6,11 +6,12 @@ with safe_import_context() as import_ctx:
     import numpy as np
     import scipy
 
+
 class Solver(BaseSolver):
     name = 'spams'
 
     install_cmd = 'conda'
-    requirements = ['numpy','mkl','pip:spams']
+    requirements = ['numpy', 'mkl', 'pip:spams']
 
     def set_objective(self, X, y, lmbd):
         self.X, self.y, self.lmbd = X, y, lmbd
@@ -21,18 +22,19 @@ class Solver(BaseSolver):
             self.X = np.asfortranarray(self.X)
 
         self.solver_parameter = dict(
-            lambda1 = self.lmbd, verbose = True
+            lambda1=self.lmbd, verbose=True
         )
 
     def run(self, n_iter):
-        y=np.expand_dims(np.asfortranarray(self.y), axis = 1)
+        y = np.expand_dims(np.asfortranarray(self.y), axis=1)
         if (scipy.sparse.issparse(self.X)):
-            W0 = np.zeros((self.X.shape[1],1), dtype = y.dtype, order = "F")
-            self.w = fistaFlat(y,self.X,W0, **self.solver_parameter, regul = 'l1',
-                    it0 = 10000, loss = 'square', tol = 1e-12, max_it = n_iter).flatten()
+            W0 = np.zeros((self.X.shape[1], 1), dtype=y.dtype, order="F")
+            self.w = fistaFlat(y, self.X, W0, **self.solver_parameter,
+                               regul='l1', it0=10000, loss='square', tol=1e-12,
+                               max_it=n_iter).flatten()
         else:
-            self.w = lasso(y,  D = self.X, L = n_iter,
-                        **self.solver_parameter).toarray().flatten()
+            self.w = lasso(y,  D=self.X, L=n_iter,
+                           **self.solver_parameter).toarray().flatten()
 
     def get_result(self):
         return self.w
